move tests out of `Example` submodule
