+++ shelley/chain-and-ledger/executable-spec/shelley-spec-ledger.cabal
@@ -114,41 +114,41 @@ test-suite shelley-spec-ledger-test
     type:                exitcode-stdio-1.0
     main-is:             Tests.hs
     other-modules:       Test.Shelley.Spec.Ledger.Utils
-                         Test.Shelley.Spec.Ledger.ConcreteCryptoTypes
                          Test.Cardano.Crypto.VRF.Fake
                          Test.Shelley.Spec.Ledger.Address
-                         Test.Shelley.Spec.Ledger.Generator.Constants
+                         Test.Shelley.Spec.Ledger.CDDL
+                         Test.Shelley.Spec.Ledger.ConcreteCryptoTypes
+                         Test.Shelley.Spec.Ledger.Examples
+                         Test.Shelley.Spec.Ledger.Fees
                          Test.Shelley.Spec.Ledger.Generator.Block
+                         Test.Shelley.Spec.Ledger.Generator.Constants
                          Test.Shelley.Spec.Ledger.Generator.Core
-                         Test.Shelley.Spec.Ledger.Generator.Trace.Chain
-                         Test.Shelley.Spec.Ledger.Generator.Trace.Ledger
-                         Test.Shelley.Spec.Ledger.Generator.Trace.DCert
                          Test.Shelley.Spec.Ledger.Generator.Delegation
                          Test.Shelley.Spec.Ledger.Generator.Presets
+                         Test.Shelley.Spec.Ledger.Generator.Trace.Chain
+                         Test.Shelley.Spec.Ledger.Generator.Trace.DCert
+                         Test.Shelley.Spec.Ledger.Generator.Trace.Ledger
                          Test.Shelley.Spec.Ledger.Generator.Update
                          Test.Shelley.Spec.Ledger.Generator.Utxo
+                         Test.Shelley.Spec.Ledger.MultiSigExamples
                          Test.Shelley.Spec.Ledger.NonTraceProperties.Generator
                          Test.Shelley.Spec.Ledger.NonTraceProperties.Mutator
                          Test.Shelley.Spec.Ledger.NonTraceProperties.PropertyTests
                          Test.Shelley.Spec.Ledger.PropertyTests
-                         Test.Shelley.Spec.Ledger.Shrinkers
-                         Test.Shelley.Spec.Ledger.Examples.STSTests
-                         Test.Shelley.Spec.Ledger.Examples.Examples
-                         Test.Shelley.Spec.Ledger.Examples.MultiSigExamples
-                         Test.Shelley.Spec.Ledger.Examples.UnitTests
                          Test.Shelley.Spec.Ledger.Rules.ClassifyTraces
                          Test.Shelley.Spec.Ledger.Rules.TestChain
                          Test.Shelley.Spec.Ledger.Rules.TestDeleg
+                         Test.Shelley.Spec.Ledger.Rules.TestDelegs
                          Test.Shelley.Spec.Ledger.Rules.TestLedger
                          Test.Shelley.Spec.Ledger.Rules.TestNewEpoch
-                         Test.Shelley.Spec.Ledger.Rules.TestDelegs
                          Test.Shelley.Spec.Ledger.Rules.TestPool
                          Test.Shelley.Spec.Ledger.Rules.TestPoolreap
                          Test.Shelley.Spec.Ledger.Rules.TestUtxo
                          Test.Shelley.Spec.Ledger.Rules.TestUtxow
-                         Test.Shelley.Spec.Ledger.Examples.Serialization
-                         Test.Shelley.Spec.Ledger.Examples.CDDL
-                         Test.Shelley.Spec.Ledger.Examples.Fees
+                         Test.Shelley.Spec.Ledger.Serialization
+                         Test.Shelley.Spec.Ledger.Shrinkers
+                         Test.Shelley.Spec.Ledger.STSTests
+                         Test.Shelley.Spec.Ledger.UnitTests
                          Test.TestScenario
     hs-source-dirs:      test
     ghc-options:
