+++ shelley/chain-and-ledger/executable-spec/test/Tests.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE LambdaCase #-}
 
 import Test.Shelley.Spec.Ledger.Address(addressTests)
-import Test.Shelley.Spec.Ledger.Examples.CDDL (cddlTests)
-import Test.Shelley.Spec.Ledger.Examples.STSTests (stsTests)
-import Test.Shelley.Spec.Ledger.Examples.Serialization (serializationTests)
-import Test.Shelley.Spec.Ledger.Examples.UnitTests (unitTests)
+import Test.Shelley.Spec.Ledger.CDDL (cddlTests)
+import Test.Shelley.Spec.Ledger.STSTests (stsTests)
+import Test.Shelley.Spec.Ledger.Serialization (serializationTests)
+import Test.Shelley.Spec.Ledger.UnitTests (unitTests)
 import Test.Shelley.Spec.Ledger.PropertyTests (minimalPropertyTests, propertyTests)
 import Test.Shelley.Spec.Ledger.NonTraceProperties.PropertyTests (nonTracePropertyTests)
 import Test.Tasty
