+++ shelley/chain-and-ledger/executable-spec/test/shelley-spec-ledger-test.cabal
@@ -18,9 +18,7 @@ flag development
 library
   exposed-modules:   Test.Cardano.Crypto.VRF.Fake
                      Test.Shelley.Spec.Ledger.ConcreteCryptoTypes
-                     Test.Shelley.Spec.Ledger.Examples.Examples
-                     Test.Shelley.Spec.Ledger.Examples.MultiSigExamples
-                     Test.Shelley.Spec.Ledger.Examples.Serialization
+                     Test.Shelley.Spec.Ledger.Examples
                      Test.Shelley.Spec.Ledger.Generator.Block
                      Test.Shelley.Spec.Ledger.Generator.Constants
                      Test.Shelley.Spec.Ledger.Generator.Core
@@ -31,8 +29,10 @@ library
                      Test.Shelley.Spec.Ledger.Generator.Trace.Ledger
                      Test.Shelley.Spec.Ledger.Generator.Update
                      Test.Shelley.Spec.Ledger.Generator.Utxo
-                     Test.Shelley.Spec.Ledger.PreSTSGenerator
-                     Test.Shelley.Spec.Ledger.PreSTSMutator
+                     Test.Shelley.Spec.Ledger.MultiSigExamples
+                     Test.Shelley.Spec.Ledger.NonTraceProperties.Generator
+                     Test.Shelley.Spec.Ledger.NonTraceProperties.Mutator
+                     Test.Shelley.Spec.Ledger.Serialization
                      Test.Shelley.Spec.Ledger.Shrinkers
                      Test.Shelley.Spec.Ledger.Utils
   ghc-options:
